% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_log.R
\name{get_from_log}
\alias{get_from_log}
\title{get messages for a single search string from TeX log file}
\usage{
get_from_log(texlog, pattern = "", line = grep(pattern, texlog, ignore.case
  = TRUE), nlines = 1, ncharmin = 4)
}
\arguments{
\item{texlog}{Character string vector e.g. from \code{\link{readLines}}}

\item{pattern}{Char: pattern to be searched for. DEFAULT: ""}

\item{line}{Integer: line number(s) to be returned.
DEFAULT: \code{grep(pattern, texlog, ignore.case=TRUE)}}

\item{nlines}{Integer: Number of lines to be used, starting at \code{line}.
DEFAULT: 1}

\item{ncharmin}{Integer: Lines with fewer characters are omitted from output.
DEFAULT: 4}
}
\value{
Vector of character strings. For each line matching some pattern,
        the \code{nlines} following this line number are aggregated with
        \code{\link{paste}(lines_with_content, collapse="\\n")}.
        If nothing matches, NULL is returned.
}
\description{
get lines from log file content for a single search string.
             This is a basic function for \code{\link{get_texlog_warnings}}
}
\examples{
log <- system.file("extdata/no_messages.log", package="knitPres")
log <- readLines(log)
cat(get_from_log(log, "plot_indo2-1.pdf  used", nlines=4))

}
\seealso{
\code{\link{get_texlog_warnings}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017
}
\keyword{file}
