% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_texlog_warnings.R
\name{get_texlog_warnings}
\alias{get_texlog_warnings}
\title{get messages from TeX log file}
\usage{
get_texlog_warnings(texlogfile, unique = TRUE, nlines_error = 4,
  nlines_box = 7, nlines_warning = 3, ...)
}
\arguments{
\item{texlogfile}{Char: file name}

\item{unique}{Logical. Should ...TODO... DEFAULT: TRUE}

\item{nlines_error}{Integer: number of lines for an error. DEFAULT:4}

\item{nlines_box}{Integer: number of lines for over/underfull box messages. DEFAULT:7}

\item{nlines_warning}{Integer: number of lines for other messages. DEFAULT:3 (may not always be enough)}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
Vector of character strings, one per warning
}
\description{
get messages (notes, warnings, errors) from sometexdoc.log file
}
\examples{

testlog <- function(logname, ...)
  {
  logf <- system.file(paste0("extdata/",logname,".log"), package="knitPres")
  cat("", get_texlog_warnings(logf, ...), sep="\\n---\\n")
  }

testlog("no_messages")
testlog("blitzaerror")
testlog("lotsoferrors", nlines_error=3)
testlog("lotsoferrors", nlines_error=4) # more info, less duplicates
testlog("pgferror")
testlog("superseterror")
testlog("superseterror", nlines_box=0) # no over/underful box warnings
testlog("superseterror", nlines_warning=5)

}
\references{
Test log files from \url{https://github.com/stefanhepp/pplatex/tree/master/test}
}
\seealso{
\code{\link{get_from_log}}, \code{graphics::\link[graphics]{plot}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017
}
\keyword{file}
