% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_hand_pres.R
\name{knit_hand_pres}
\alias{knit_hand_pres}
\title{knit handout and presentation version of .Rnw file}
\usage{
knit_hand_pres(file, presname = "_pres", ...)
}
\arguments{
\item{file}{Char: .Rnw file name to be knitted}

\item{presname}{Char: inset that will get appended to file name. DEFAULT: "_pres"}

\item{\dots}{Further arguments passed to \code{\link{knit_single_pres}}}
}
\value{
ReturnValue
}
\description{
knit handout and presentation version of .Rnw file in parallel
}
\examples{
\dontrun{## Exclude time consuming test from regular testing
rnwfile <- system.file("extdata/minimalpres.Rnw", package="knitPres")
dir.create("TempKnitFolder")
owd <- setwd("TempKnitFolder") ; getwd()
file.copy(rnwfile, "minimalpres.Rnw")
messages <- knit_hand_pres(file="minimalpres.Rnw")
messages
#    berryFunctions::openFile("minimalpres_pres.pdf")
setwd(owd) ; getwd()
unlink("TempKnitFolder", recursive=TRUE)
}

}
\seealso{
\code{\link{knit_single_pres}}, \code{\link{get_texlog_warnings}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017
}
\keyword{file}
