% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_single_pres.R
\name{knit_single_pres}
\alias{knit_single_pres}
\title{knit an Rnw presentation, log progress + warnings}
\usage{
knit_single_pres(file)
}
\arguments{
\item{file}{Filename}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
log messages as a single character string, invisible
}
\description{
knit an Rnw presentation, write progress + tex warnings into a
             separate .knitlog file
}
\examples{
\dontrun{## Exclude time consuming test from regular testing
rnwfile <- system.file("extdata/minimalpres.Rnw", package="knitPres")
dir.create("TempKnitFolder")
owd <- setwd("TempKnitFolder") ; getwd()
file.copy(rnwfile, "minimalpres.Rnw")
messages <- knit_single_pres(file="minimalpres.Rnw")
cat(messages)
#    berryFunctions::openFile("minimalpres.pdf")
setwd(owd) ; getwd()
unlink("TempKnitFolder", recursive=TRUE)
}

}
\seealso{
\code{\link{get_texlog_warnings}} for extracting messages from the TeX .log file
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017
}
\keyword{file}
