% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateTextbox.R
\name{locateTextbox}
\alias{locateTextbox}
\title{determine locations for textbox}
\usage{
locateTextbox(
  img = "C:/Users/berry/Desktop/pdfscreenshot.PNG",
  calibrate = FALSE,
  ...
)
}
\arguments{
\item{img}{Image file name, is read with \code{png::\link{readPNG}(img)}}

\item{calibrate}{Show template for calibration? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
invisible list with coordinates, the printed message contains ready-to-use code
}
\description{
extract locations for textbox from pdf screenshot with reference markers
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2021
}
