% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_clean.R
\name{tex_clean}
\alias{tex_clean}
\title{Clean up after tex run}
\usage{
tex_clean(file, ext = NULL)
}
\arguments{
\item{file}{Filename, with or without extension}

\item{ext}{Further extensions to also be removed, e.g. .tex/.log}
}
\value{
output from \code{\link{unlink}}
}
\description{
Clean manually after tex run. .tex + .log files are kept.
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2021
}
