% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texlog_warnings.R
\name{texlog_warnings}
\alias{texlog_warnings}
\title{Display tex log warnings}
\usage{
texlog_warnings(
  file,
  nlines_error = 4,
  nlines_box = 7,
  nlines_warning = 3,
  ...
)
}
\arguments{
\item{file}{Char: file name (.Rnw, .tex or .log)}

\item{nlines_error}{Integer: number of lines for an error. DEFAULT: 4}

\item{nlines_box}{Integer: number of lines for over/underfull box messages. DEFAULT: 7}

\item{nlines_warning}{Integer: number of lines for other messages. DEFAULT: 3 (may not always be enough)}

\item{\dots}{Further arguments passed to \code{\link{readLines}}}
}
\value{
Vector of character strings, one per warning
}
\description{
Display (if any are found) notes, warnings and errors from some_latex_file.log
}
\examples{
testlog <- function(logname, ...)
  {
  logf <- system.file(paste0("extdata/",logname,".log"), package="knitPres")
  texlog_warnings(logf, ...)
  }
testlog("no_messages")
testlog("blitzaerror")
testlog("lotsoferrors", nlines_error=3)
testlog("lotsoferrors", nlines_error=4) # more info, less duplicates
testlog("pgferror")
testlog("superseterror")
testlog("superseterror", nlines_box=0) # no over/underful box warnings
testlog("superseterror", nlines_warning=5)
testlog("hashtagURL") # not formatted as error!!

}
\references{
Test log files from \url{https://github.com/stefanhepp/pplatex/tree/master/test}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017 + Jul 2021
}
\keyword{file}
