% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_hand_pres.R
\name{knit_hand_pres}
\alias{knit_hand_pres}
\title{knit handout and presentation version of .Rnw file}
\usage{
knit_hand_pres(file, presname = "_pres", open = TRUE, cleanup = TRUE, ...)
}
\arguments{
\item{file}{Char: .Rnw file name to be knitted}

\item{presname}{Char: inset that will get appended to file name. DEFAULT: "_pres"}

\item{open}{Logical: Open the resulting pdfs with
\code{berryFunctions::\link{openFile}}? DEFAULT: TRUE}

\item{cleanup}{Logical: remove intermediate tex files? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{tools::\link{texi2pdf}}}
}
\value{
log entries obtained with \code{\link{get_texlog_warnings}}
}
\description{
knit handout and presentation version of .Rnw file,
             compile tex file in parallel,
             write progress + tex warnings into a separate .knitlog file
}
\examples{
\dontrun{## Exclude time consuming test from regular testing

# Copy Rnw file to temporary folder ---
rnwfile <- system.file("extdata/minimalpres.Rnw", package="knitPres")
dir.create("TempKnitFolder")
owd <- setwd("TempKnitFolder") ; getwd()
file.copy(rnwfile, "minimalpres.Rnw")

# Actual knitting ----
knit_hand_pres(file="minimalpres.Rnw")

# clean up ----
setwd(owd) ; getwd()
unlink("TempKnitFolder", recursive=TRUE)
}

}
\seealso{
\code{\link{texlog_warnings}} for extracting messages from the TeX .log file
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2017
}
\keyword{file}
