% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateTextbox.R
\name{locateTextbox}
\alias{locateTextbox}
\title{Determine locations for textbox}
\usage{
locateTextbox(
  img = "C:/Users/berry/Desktop/pdfscreenshot.PNG",
  frame = TRUE,
  ...
)
}
\arguments{
\item{img}{Image file name, is read with \code{png::\link{readPNG}(img)}
DEFAULT: "C:/Users/berry/Desktop/pdfscreenshot.PNG"}

\item{frame}{Show complete frame template? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{plot}}}
}
\value{
Invisible list with coordinates, the printed message contains ready-to-use code
}
\description{
Extract locations for textbox from pdf screenshot with reference markers
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2021
}
